+++ .github/workflows/ci.yml
@@ -28,40 +28,6 @@ jobs:
       - name: Run ESLint
         run: npm run lint -- --max-warnings 250
 
-  migration-tests:
-    name: Migration Tests
-    runs-on: ubuntu-latest
-    needs: lint-typecheck
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Setup Supabase CLI
-        uses: supabase/setup-cli@v1
-        with:
-          version: latest
-
-      - name: Start Supabase local services
-        run: supabase start
-
-      - name: Wait for services to be healthy
-        run: |
-          echo "Waiting for Supabase services to be healthy..."
-          timeout 60 bash -c 'until supabase status | grep -q "API URL"; do sleep 2; done'
-          echo "Services are ready!"
-
-      - name: Run migration validation
-        run: ./scripts/validate-migrations.sh
-
-      - name: Run database reset
-        run: supabase db reset
-
-      - name: Run smoke tests
-        run: ./scripts/test-migrations.sh
-
-      - name: Cleanup
-        if: always()
-        run: supabase stop --no-backup
-
   unit-tests:
     name: Unit Tests
     runs-on: ubuntu-latest
@@ -102,48 +68,3 @@ jobs:
         env:
           VITE_SUPABASE_URL: https://example.supabase.co
           VITE_SUPABASE_ANON_KEY: placeholder-for-build
-
-      - name: Upload build artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: dist
-          path: dist/
-          retention-days: 7
-
-  e2e-tests:
-    name: E2E Tests
-    runs-on: ubuntu-latest
-    needs: build
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: ${{ env.NODE_VERSION }}
-          cache: 'npm'
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Install Playwright browsers
-        run: npx playwright install --with-deps chromium
-
-      - name: Download build artifacts
-        uses: actions/download-artifact@v4
-        with:
-          name: dist
-          path: dist/
-
-      - name: Run E2E tests
-        run: npm run test:e2e
-        env:
-          CI: true
-
-      - name: Upload Playwright report
-        uses: actions/upload-artifact@v4
-        if: always()
-        with:
-          name: playwright-report
-          path: playwright-report/
-          retention-days: 30
